backend/requirements.txt
Flask==2.3.3
mysql-connector-python==8.3.0
Werkzeug==2.3.7
APScheduler==3.10.4
python-dotenv==1.0.0
pymysql
cryptography
gunicorn
